ManyLinux Wheel Building workflow (#49)
Seems like there are a number of issues related to installing this package.  Likely related to hdf installing its libs in slightly non-standard locations.  `/usr/lib64/hdf/` instead of `/usr/lib64/`.  Conda solves a lot of those problems but pip wheels are easier in containers and a nice alternative.  This PR would create self contained wheels for pip installation.  

Here's a relatively simple GitHub Actions workflow which would build (and potentially upload) a variety of static linked manylinux wheels to PyPI.  This removes the per-requisite of installing `hdf` manually as the libraries are included in the wheels.  

Wheels are currently uploaded to the artifact section under the CI action run, but the triggers can be changed to only run on a release tag and on master branch.   Running workflow and artifact from this branch here: https://github.com/abrammer/pyhdf/actions/runs/775429130

Other thoughts:
Users could retain the current installation method with the `--no-binary`  flag on pip. 
HDF license allows distribution within downstream packages as far as I can tell, and the built wheels are relatively small ~600kb. 

TODO:
 - [x] update installation docs if idea is accepted
 - [x] update licence to make sure the included HDF code is mentioned, whatever bsd needs (just a mention I think)
 - [x] pypi secrets added to repo for auto deploy.  or the wheels could just be manually pushed from the artifact upload. 
 - [x] change the repo url in the pypa action to production pypi